Reinstate SQLServerPlatform and SQLServerKeywords
