+++ psalm.xml.dist
@@ -65,7 +65,9 @@
                     TODO: remove in 4.0.0
                 -->
                 <referencedClass name="Doctrine\DBAL\Platforms\Keywords\PostgreSQL94Keywords"/>
+                <referencedClass name="Doctrine\DBAL\Platforms\Keywords\SQLServer2012Keywords"/>
                 <referencedClass name="Doctrine\DBAL\Platforms\PostgreSQL94Platform"/>
+                <referencedClass name="Doctrine\DBAL\Platforms\SQLServer2012Platform"/>
             </errorLevel>
         </DeprecatedClass>
         <DeprecatedInterface>
@@ -182,7 +184,7 @@
                 <file name="src/DriverManager.php"/>
                 <file name="src/Platforms/AbstractPlatform.php"/>
                 <file name="src/Platforms/MySQLPlatform.php"/>
-                <file name="src/Platforms/SQLServer2012Platform.php"/>
+                <file name="src/Platforms/SQLServerPlatform.php"/>
                 <file name="src/Platforms/SqlitePlatform.php"/>
                 <file name="src/Schema/Column.php"/>
                 <!--
