+++ tests/Driver/AbstractSQLServerDriverTest.php
@@ -8,11 +8,12 @@ use Doctrine\DBAL\Driver\API\ExceptionConverter as ExceptionConverterInterface;
 use Doctrine\DBAL\Driver\API\SQLSrv\ExceptionConverter;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
 use Doctrine\DBAL\Platforms\SQLServer2012Platform;
+use Doctrine\DBAL\Platforms\SQLServerPlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\SQLServerSchemaManager;
 
 /**
- * @extends AbstractDriverTest<SQLServer2012Platform>
+ * @extends AbstractDriverTest<SQLServerPlatform>
  */
 abstract class AbstractSQLServerDriverTest extends AbstractDriverTest
 {
