+++ tests/Functional/LockMode/NoneTest.php
@@ -9,7 +9,7 @@ use Doctrine\DBAL\Driver\OCI8;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\LockMode;
 use Doctrine\DBAL\Platforms\SqlitePlatform;
-use Doctrine\DBAL\Platforms\SQLServer2012Platform;
+use Doctrine\DBAL\Platforms\SQLServerPlatform;
 use Doctrine\DBAL\Schema\Table;
 use Doctrine\DBAL\Tests\FunctionalTestCase;
 use Doctrine\DBAL\Tests\TestUtil;
@@ -29,7 +29,7 @@ class NoneTest extends FunctionalTestCase
             self::markTestSkipped('This test fails on OCI8 for a currently unknown reason');
         }
 
-        if ($this->connection->getDatabasePlatform() instanceof SQLServer2012Platform) {
+        if ($this->connection->getDatabasePlatform() instanceof SQLServerPlatform) {
             // Use row versioning instead of locking on SQL Server (if we don't, the second connection will block when
             // attempting to read the row created by the first connection, instead of reading the previous version);
             // for some reason we cannot set READ_COMMITTED_SNAPSHOT ON when not running this test in isolation,
@@ -71,7 +71,7 @@ class NoneTest extends FunctionalTestCase
 
         $this->connection->getSchemaManager()->dropTable('users');
 
-        if (! $this->connection->getDatabasePlatform() instanceof SQLServer2012Platform) {
+        if (! $this->connection->getDatabasePlatform() instanceof SQLServerPlatform) {
             return;
         }
